refactor(Annotation): standardize components containers names
