+++ src/Ml.Cli.WebApp/ClientApp/src/Annotation/Toolkit/TagOverText/TagOverTextLazy.js
@@ -2,7 +2,7 @@
 
 const TagOverTextLazy = ({expectedOutput, url, onSubmit}) => {
     
-    const TagOverTextContainer = React.lazy(() => import('./TagOverTextContainer'));
+    const TagOverTextContainer = React.lazy(() => import('./TagOverText.container'));
     
     return <Suspense fallback={<div>Chargement de TagOverText...</div>}>
         <TagOverTextContainer
